// src/config/dimmers.d.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export type LightFrLumiere = "light";
export type FanFrVentilation = "fan";
export type WindowCoveringFrVoletRideau = "covering";
export type Dimmers = {
  displayName: string;
  type: (LightFrLumiere | FanFrVentilation | WindowCoveringFrVoletRideau) & string;
  index: number;
  anaIndex: number;

};